% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brcamutSE}
\alias{brcamutSE}
\title{a SummarizedExperiment collecting BRCA mutations and colData from curatedTCGAData}
\format{
SummarizedExperiment
}
\usage{
brcamutSE
}
\description{
a SummarizedExperiment collecting BRCA mutations and colData from curatedTCGAData
}
\keyword{datasets}
