% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survcode.R
\name{build_simple_survival_curve}
\alias{build_simple_survival_curve}
\title{utility for survival curve illustration}
\usage{
build_simple_survival_curve()
}
\description{
utility for survival curve illustration
}
