% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_surv.R
\name{build_surv_for_mut}
\alias{build_surv_for_mut}
\title{use curatedTCGAData to produce survival time structure and mutation matrix for a TCGA tumor type}
\usage{
build_surv_for_mut(project = "BRCA", min.numevents = 10)
}
\arguments{
\item{project}{character(1), defaults to "BRCA"}

\item{min.numevents}{numeric(1), defaults to 10; fail if there are fewer than this number of events overall}
}
\value{
a list with components surv, coldata, and mutdata
}
\description{
use curatedTCGAData to produce survival time structure and mutation matrix for a TCGA tumor type
}
\note{
observations lacking positive follow up time are silently omitted
}
\examples{
requireNamespace("survival")
br = build_surv_for_mut("BRCA")
has_TTN = apply(br$mutdata, 2, function(x) any(x == "TTN", na.rm=TRUE))
fi = survival::survfit(br$surv ~ has_TTN)
plot(fi, lwd=2, col=c("blue", "orange"), xlab = "t = Years from diagnosis", ylab="S(t) = Prob(survive beyond t)")  # KM-plot
legend(.1, .3, lwd=2, col=c("blue", "orange"), lty=1, legend=c("TTN wild-type", "TTN-mutant"), bty="n")
title("TCGA BRCA survival")
}
