% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canada_crude}
\alias{canada_crude}
\title{a data.frame with (rounded) crude rates of cancer as presented at https://www.statcan.gc.ca/en/dai/btd/asr}
\format{
data.frame
}
\usage{
canada_crude
}
\description{
a data.frame with (rounded) crude rates of cancer as presented at https://www.statcan.gc.ca/en/dai/btd/asr
}
\keyword{datasets}
