% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer_map_world.R
\name{cancer_map_world}
\alias{cancer_map_world}
\title{simple map of cancer rates world wide}
\usage{
cancer_map_world(world_tab = world_cancer_2020_all)
}
\arguments{
\item{world_tab}{a data.frame as produced by gco.iarc.fr}
}
\description{
simple map of cancer rates world wide
}
\note{
used https://gco.iarc.fr/today/online-analysis-map table CSV exports, some
cleanup of special characters
}
\examples{
data(world_cancer_2020_all)
cancer_map_world()
}
