% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_surv.R
\name{compare_tumors}
\alias{compare_tumors}
\title{compare multiple tumor times with respect to survival}
\usage{
compare_tumors(types = c("BRCA", "GBM"), ...)
}
\arguments{
\item{types}{character() TCGA project codes}

\item{\dots}{passed to `build_surv_for_mut`}
}
\value{
a list of lists of two elements: survlist (Surv objects) and
mutlist (mutation matrices)
}
\description{
compare multiple tumor times with respect to survival
}
\examples{
cmp = compare_tumors()
plot(survfit(cmp$survlist[[1]]~1), conf.int=FALSE, xlab="years", ylab="Survival prob.", main="TCGA survival", lwd=2)
lines(survfit(cmp$survlist[[2]]~1), lwd=2, lty=2, conf.int=FALSE)
legend(15, .98, lty=c(1,2), lwd=2, legend=c("BRCA", "GBM"), bty="n")
}
