% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{filter_woncan}
\alias{filter_woncan}
\title{generate site-specific rate table}
\usage{
filter_woncan(site)
}
\arguments{
\item{site}{character(1) name of anatomic site, must reside in value of `woncan_types()`}
}
\description{
generate site-specific rate table
}
