% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_hcc.R
\name{get_hcc_spatial_path}
\alias{get_hcc_spatial_path}
\title{retrieve or cache SingleCellExperiment data for spatial transcriptomics of HCC}
\usage{
get_hcc_spatial_path(sample = "1R", cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{sample}{character(1) either "1R" or "6NR"}

\item{cache}{cache instance likely inheriting from BiocFileCache::BiocFileCache}
}
\value{
character(1) path to rda file
}
\description{
retrieve or cache SingleCellExperiment data for spatial transcriptomics of HCC
}
\note{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSM7661255} is
the basis for the "1R" example.
\url{https://pubmed.ncbi.nlm.nih.gov/37723590/} is the primary paper.
}
\examples{
get_hcc_spatial_path()
}
