% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_stuff.R
\name{has_mut_in}
\alias{has_mut_in}
\title{check all columns for presence of symbol in a SE with ragged mutation matrix as assay}
\usage{
has_mut_in(sym, se)
}
\arguments{
\item{sym}{character(1) HUGO symbol}

\item{se}{a SummarizedExperiment instance with ragged mutation matrix as assay}
}
\value{
a vector of logicals, one for each column; uses `%in%`
}
\description{
check all columns for presence of symbol in a SE with ragged mutation matrix as assay
}
\examples{
data(brcamutSE)
table(has_mut_in("TTN", brcamutSE))
}
