% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{make_comparison}
\alias{make_comparison}
\title{make an interactive histogram of rate ratios, or a scatterplot for rate comparisons, between two areas}
\usage{
make_comparison(
  MSA1 = "Baton Rouge, LA",
  MSA2 = "Tucson, AZ",
  logscale = TRUE,
  type = "scatter",
  min.rate = 10,
  rate_table
)
}
\arguments{
\item{MSA1}{character(1) element of column 'MSA' that must be present in `rate_table`}

\item{MSA2}{character(1) element of column 'MSA' that must be present in `rate_table`}

\item{logscale}{logical(1) only for type "scatter"; if TRUE (default) log10 transformation is applied to rates}

\item{type}{character(1) one of "scatter" or "ratios"; if the latter, a histogram of ratios is produced with MSA1 rates in numerator}

\item{min.rate}{numeric(1) minimum number per 100000 for cancer type to be included}

\item{rate_table}{data.frame like `woncan` in YESCDS or BiocYES}
}
\description{
make an interactive histogram of rate ratios, or a scatterplot for rate comparisons, between two areas
}
\examples{
data(woncan)
make_comparison(type="scatter", rate_table=woncan)
make_comparison(type="ratios", rate_table=woncan)
}
