% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{make_hist}
\alias{make_hist}
\title{make an interactive histogram for a given cancer site, based on a rate table}
\usage{
make_hist(site = "Prostate", binwidth = 5, rate_table)
}
\arguments{
\item{site}{character(1) element of column 'Cancer Sites' that must be present in `rate_table`}

\item{binwidth}{numeric(1) passed to `ggplot2::geom_histogram`}

\item{rate_table}{data.frame like `woncan` in YESCDS or BiocYES}
}
\description{
make an interactive histogram for a given cancer site, based on a rate table
}
\examples{
data(woncan)
make_hist(rate_table=woncan)
}
