% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{make_pair}
\alias{make_pair}
\title{make interactive beeswarm or box plots for all cancer incidence rates in two areas}
\usage{
make_pair(
  MSA1 = "Baton Rouge, LA",
  MSA2 = "Tucson, AZ",
  type = "boxplot",
  rate_table
)
}
\arguments{
\item{MSA1}{character(1) element of column 'MSA' that must be present in `rate_table`}

\item{MSA2}{character(1) element of column 'MSA' that must be present in `rate_table`}

\item{type}{character(1) either "boxplot" or "beeswarm"}

\item{rate_table}{data.frame like `woncan` in YESCDS or BiocYES}
}
\description{
make interactive beeswarm or box plots for all cancer incidence rates in two areas
}
\examples{
data(woncan)
make_hist(rate_table=woncan)
}
