% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_map.R
\name{mass_cancer_map}
\alias{mass_cancer_map}
\title{map cancer rates for selected sites to MA counties}
\usage{
mass_cancer_map(site = "breast")
}
\arguments{
\item{body_site}{character(1) defaults to "breast", can also take value "prostate"}
}
\description{
map cancer rates for selected sites to MA counties
}
\examples{
mass_cancer_map()
}
