% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metabric}
\alias{metabric}
\title{a data.frame with clinical features of a breast cancer study (METABRIC) derived from cBioPortal}
\format{
data.frame
}
\usage{
metabric
}
\description{
a data.frame with clinical features of a breast cancer study (METABRIC) derived from cBioPortal
}
\keyword{datasets}
