% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered_seg_cal.R
\name{ordered_seg_cal}
\alias{ordered_seg_cal}
\title{big display of ordered CIs for California Lung and Bronchus incidence 2014-2018}
\usage{
ordered_seg_cal()
}
\description{
big display of ordered CIs for California Lung and Bronchus incidence 2014-2018
}
