% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seer_trend.R
\name{plot_seer_trend}
\alias{plot_seer_trend}
\title{plot the rate statistics over time for selected cancer sites and subgroups}
\usage{
plot_seer_trend(
  tab,
  site.name = "Lung and Bronchus",
  gender = "Male",
  race = "All Races"
)
}
\arguments{
\item{tab}{a data.frame like that obtained with `data(SEER_2018)`}

\item{site.name}{character(1) cancer site name as used by SEER, possible choices are
`"Lung and Bronchus", "Pancreas", "Leukemia", "Kidney and Renal Pelvis"`}

\item{gender}{character(1) either "Male" or "Female"}

\item{race}{character(1) can be "All Races", or "White", or "Black"}
}
\description{
plot the rate statistics over time for selected cancer sites and subgroups
}
\examples{
data(SEER_2018)
plot_seer_trend(SEER_2018, site.name = "Kidney and Renal Pelvis")
}
