% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIplots.R
\name{plotwci}
\alias{plotwci}
\title{simple display of confidence intervals for cancer incidence}
\usage{
plotwci(
  tab,
  metric = "incidence per 100000 2014-2018",
  site = "Overall",
  xlab = "Massachusetts counties",
  ylim = c(350, 620)
)
}
\arguments{
\item{tab}{a CSV file with columns as provided via \url{https://gis.cdc.gov/CANCER/USCS/#/StateCounty/}}

\item{metric}{character(1) detailed description of statistic}

\item{site}{character(1) name of anatomic site}

\item{xlab}{character(1) label for x axis}

\item{ylim}{numeric(2) extent of y axis in units of incidence per 100000}
}
\description{
simple display of confidence intervals for cancer incidence
}
\examples{
oldpar = par(no.readonly=TRUE)
prost = read.csv(system.file("csv/MASSProstateWCI.csv", package="YESCDS"))
bre = read.csv(system.file("csv/MABreastWCI.csv", package="YESCDS"))
par(mfrow=c(1,2))
plotwci(prost, site="Prostate", ylim=c(60,270))
plotwci(bre, site="Breast", ylim=c(60,280))
par(oldpar)
}
