% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prostate_met_adeno}
\alias{prostate_met_adeno}
\title{a data.frame with clinical features of a metastatic prostate adenocarcinoma study}
\format{
data.frame
}
\usage{
prostate_met_adeno
}
\description{
a data.frame with clinical features of a metastatic prostate adenocarcinoma study
}
\keyword{datasets}
