% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SEER_2018}
\alias{SEER_2018}
\title{a data.frame with SEER estimates of cancer incidence for 2018}
\format{
data.frame
}
\usage{
SEER_2018
}
\description{
a data.frame with SEER estimates of cancer incidence for 2018
}
\note{
Collected at \url{https://seer.cancer.gov/statistics/preliminary-estimates/data-downloads/Feb-2019-Preliminary-Estimates.xlsx} on May 27 2022.  Notes include: "The estimates in this file are preliminary estimates for the November 2019 SEER submission based on the February 2019 SEER submission. These rates are subject to change prior to the official SEER release of estimates in April 2020.
Source: SEER 18 areas (San Francisco, Connecticut, Detroit, Hawaii, Iowa, New Mexico, Seattle, Utah, Atlanta, San Jose-Monterey, Los Angeles, Alaska Native Registry, Rural Georgia, California excluding SF/SJM/LA, Kentucky, Louisiana, New Jersey and Georgia excluding ATL/RG).
Rates are per 100,000 and are age-adjusted to the 2000 US Std Population (19 age groups - Census P25-1130).
}
\keyword{datasets}
