% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survcode.R
\name{show_5y_estimate}
\alias{show_5y_estimate}
\title{utility for survival curve illustration: 5y survprob estimate sketch}
\usage{
show_5y_estimate()
}
\description{
utility for survival curve illustration: 5y survprob estimate sketch
}
