% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survcode.R
\name{show_median_estimate}
\alias{show_median_estimate}
\title{utility for survival curve illustration: median estimate sketch}
\usage{
show_median_estimate()
}
\description{
utility for survival curve illustration: median estimate sketch
}
