% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statcrunch_msa}
\alias{statcrunch_msa}
\title{geographic information from statcrunch.com on metropolitan statistical areas}
\format{
data.frame
}
\usage{
statcrunch_msa
}
\description{
geographic information from statcrunch.com on metropolitan statistical areas
}
\keyword{datasets}
