% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{table_woncan}
\alias{table_woncan}
\title{simplify table view of WONDER CDC statistsics}
\usage{
table_woncan(site)
}
\arguments{
\item{site}{character(1) name of anatomic site, must reside in value of `woncan_types()`}
}
\description{
simplify table view of WONDER CDC statistsics
}
