% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_stuff.R
\name{two_sample_surv_setup}
\alias{two_sample_surv_setup}
\title{two-sample survival comparison dependent on mutation present in gene}
\usage{
two_sample_surv_setup(sym, se, timevar = "OS.Time", eventvar = "OS.event")
}
\arguments{
\item{sym}{character(1) HUGO symbol}

\item{se}{a SummarizedExperiment instance with ragged mutation matrix as assay}

\item{timevar}{character(1) variable found in colData(se) recording time to event}

\item{eventvar}{character(1) variable found in colData(se) recording death or censoring}
}
\description{
two-sample survival comparison dependent on mutation present in gene
}
\examples{
data(brcamutSE)
ss = two_sample_surv_setup("TTN", brcamutSE)
summary(ss)
plot(ss, lty=1:2, lwd=2)
legend(500,.4,  lty=1:2, lwd=2, legend=c("TTN WT", "TTN Mut"))
}
