% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_map.R
\name{us_map}
\alias{us_map}
\title{simple function to display HTML map of US}
\usage{
us_map(provider = "CartoDB.Positron")
}
\arguments{
\item{provider}{character(1) passed to `leaflet::addProviderTiles`, defaults to
`"CartoDB.Positron"`}
}
\description{
simple function to display HTML map of US
}
\examples{
if (interactive()) mass_map()
}
