% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapp.R
\name{variation_app}
\alias{variation_app}
\title{simple app to show different ways of examining variation over body sites and MSAs}
\usage{
variation_app()
}
\description{
simple app to show different ways of examining variation over body sites and MSAs
}
