% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vjc_cancer_net}
\alias{vjc_cancer_net}
\title{a table of data of Vince's contacts who developed cancer}
\format{
data.frame
}
\usage{
vjc_cancer_net
}
\description{
a table of data of Vince's contacts who developed cancer
}
\keyword{datasets}
