% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{woncan}
\alias{woncan}
\title{cancer rate data by metropolitan statistical area, 1999-2018, from CDC WONDER system}
\format{
data.frame
}
\usage{
woncan
}
\description{
cancer rate data by metropolitan statistical area, 1999-2018, from CDC WONDER system
}
\note{
Created by joining statcrunch location data to result of WONDER request.
See woncan_meta for details.
}
\keyword{datasets}
