% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{woncan_meta}
\alias{woncan_meta}
\title{metadata on cancer rate data by metropolitan statistical area, 1999-2018, from CDC WONDER system}
\format{
character
}
\usage{
woncan_meta
}
\description{
metadata on cancer rate data by metropolitan statistical area, 1999-2018, from CDC WONDER system
}
\keyword{datasets}
