% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_cancer_2020_all}
\alias{world_cancer_2020_all}
\title{a table of cancer rate data by country, retrieved from https://gco.iarc.fr/today/online-analysis-map}
\format{
data.frame
}
\usage{
world_cancer_2020_all
}
\description{
a table of cancer rate data by country, retrieved from https://gco.iarc.fr/today/online-analysis-map
}
\keyword{datasets}
