% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_geo_sf}
\alias{world_geo_sf}
\title{a simple features (sf) instance with boundaries and centroids for 241 countries or other political geographic entities}
\format{
sf
}
\usage{
world_geo_sf
}
\description{
a simple features (sf) instance with boundaries and centroids for 241 countries or other political geographic entities
}
\keyword{datasets}
